+++ Vampire Survivors Final Version/Game Loop/Shooter.h
@@ -13,7 +13,7 @@ class Shooter : public Enemy
 private:
 
 protected:
-	float health;
+	int health;
 
 public:
 	bool enemyAlive = true;
